+++ src/main.tsx
@@ -1,9 +1,28 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './App.tsx'
+import { RouterProvider, createBrowserRouter } from 'react-router-dom'
+import Root from './routes/root.tsx'
+import ErrorPage from './routes/ErrorPage.tsx'
+import HomePage from './routes/HomePage.tsx'
+import GlobalStyles from './styles/globalStyles.tsx'
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <Root />,
+    errorElement: <ErrorPage />,
+    children: [
+      {
+        element: <HomePage />,
+        index: true
+      }
+    ]
+  }
+]);
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
-    <App />
+    <GlobalStyles />
+    <RouterProvider router={router} />
   </React.StrictMode>,
 )