+++ src/routes/ErrorPage.tsx
@@ -1,0 +1,14 @@
+import { useRouteError } from "react-router-dom";
+
+const ErrorPage = () => {
+    const error : any = useRouteError();
+
+    return (
+        <div>
+            there's been an eror
+            {error.statusText || error.message};
+        </div>
+    )
+}
+
+export default ErrorPage;