+++ CHANGELOG.md
@@ -8,3 +8,16 @@
  - Starter files for React project.
  - Assets like: fonts, images.
  - Starter CSS variables.
+
+ ## [0.0.1] - 25-04-2023
+
+ ### Added
+  - React Router setup
+  - Global Styles
+  - Header Component
+  - Section Styled Component for Section Elements
+  - Logo Component
+  - Navigation Component
+  - CartButton component
+  - Burger Component
+  - RWD for Header with Burger functionality.